refactor: create useIsHarvestable hook
