+++ hooks/index.ts
@@ -13,6 +13,7 @@ export { useDebouncedValue } from './useDebouncedValue'
 export { useDisconnect } from './useDisconnect'
 export { useFiat } from './useFiat'
 export { useHarvest } from './useHarvest'
+export { useIsHarvestable } from './useIsHarvestable'
 export { useImportToken } from './useImportToken'
 export { useIsMounted } from './useIsMounted'
 export { useIsomorphicLayoutEffect } from './useIsomorphicLayoutEffect'
